+++ firmware/f7/Core/Inc/main.h
@@ -52,7 +52,7 @@ struct Cam_config {
 	uint8_t img_format;
 	uint16_t x_res;
 	uint16_t y_res;
-	uint16_t FPS;
+	uint8_t FPS; 
 	uint16_t FB_size;
 	uint16_t FIFO_width;
 	uint8_t jpeg_comp_ratio; //1 - 63 lower = more compression
